feat: Update project name and remove logo file
This commit includes the removal of the old SVG logo file and updates where the project name appears. The project name was changed in the ReadMe.md file and in the repository-open-graph.drawio file to "JoFo.NetSetUp.WorkerService" to better reflect the intention of the project. The SVG logo was removed as it's no longer relevant to the updated project.